Summary: Info about number of trainable tokens, allow floats for max train loss decrement, git ignore old models

+++ model_data_definition.py
@@ -20,7 +20,7 @@ class ModelDataDefinition:
 
             # Read settings
             self.max_train_seconds = int( ModelDataDefinition._read_setting( json_metadata , 'MaxTrainSeconds' , '0' ) )
-            self.min_loss_percentage = int( ModelDataDefinition._read_setting( json_metadata , 'MinLossPercentage' , '0' ) )
+            self.min_loss_percentage = float( ModelDataDefinition._read_setting( json_metadata , 'MinLossPercentage' , '0' ) )
             self.max_epochs = ModelDataDefinition._read_setting( json_metadata , 'MaxEpochs' , '10' )
             self.sequence_length = int( ModelDataDefinition._read_setting( json_metadata , 'SequenceLength' , '128' ) )
             self.trainable_column_index = int( ModelDataDefinition._read_setting( json_metadata , 'TrainableColumnIndex' , '-1' ) )
